bgum tests: auto add proofs to instructions
The testing framework now automatically adds proof as additional
accounts for the relevant instructions (but only if some other set
of additional accounts hasn't been already configured to generate
a certain testing scenario). Also updated tests to work with a
context that mints multiple nfts now (we could only work with one
before the ability to explicitly specify the proofs).

Signed-off-by: Alexandru Agache <alexandruag@gmail.com>